Refactor componentes front
