+++ pages/admin/interviews.tsx
@@ -5,8 +5,8 @@ import { useQuery } from '@apollo/client';
 import { GET_INTERVIEWS } from 'graphql/queries/interviews';
 import { useSession } from 'next-auth/react';
 import { Button, Dialog } from '@mui/material';
-import MakeInterviewDialog from '@components/admin/interviews/MakeInterviewDialog';
 import CreateInterviewDialog from '@components/admin/interviews/CreateInterviewDialog';
+import InterviewItem from '@components/admin/interviews/InterviewItem';
 
 export async function getServerSideProps(context) {
   return {
@@ -33,7 +33,6 @@ const Interviews = () => {
   });
 
   if (loading) return <LoadingComponent />;
-  console.log(data);
 
   return (
     <div>
@@ -56,7 +55,7 @@ const Interviews = () => {
         </div>
         <div className='flex flex-col items-center'>
           {data.getInterviewsByInterviewerId.map((i) => (
-            <Interview key={i.id} interview={i} />
+            <InterviewItem key={i.id} interview={i} />
           ))}
         </div>
       </div>
@@ -65,42 +64,6 @@ const Interviews = () => {
           closeDialog={closeDialog}
           interviwerId={interviewerId}
         />
-      </Dialog>
-    </div>
-  );
-};
-
-const Interview = ({ interview }) => {
-  const [openMakeDialog, setOpenMakeDialog] = useState(false);
-  const closeDialog = () => {
-    setOpenMakeDialog(false);
-  };
-  return (
-    <div>
-      <div className='border-2 border-inherit rounded-lg bg-slate-50 drop-shadow-lg mx-20 my-3 max-w-xl text-slate-900 hover:cursor-pointer hover:bg-slate-100'>
-        <div className='flex flex-col align-center'>
-          <h1 className='flex-col text-md font-bold text-center m-2 '>
-            {interview.name}
-          </h1>
-          <button type='button' onClick={() => setOpenMakeDialog(true)}>
-            ir
-          </button>
-          <div className='grid grid-cols-3 m-3'>
-            <div className=''>
-              <h2>{interview.admissionProcess.vacancy.position}</h2>
-              <h2>{interview.date}</h2>
-            </div>
-            <div className='flex justify-end'>
-              <h2 className='items-right m-4'>{interview.date}</h2>
-              {/* <h2 className='absolute inset-y-10 right-0 m-4'>
-                Hora: {interview.hour}
-              </h2> */}
-            </div>
-          </div>
-        </div>
-      </div>
-      <Dialog open={openMakeDialog} onClose={closeDialog}>
-        <MakeInterviewDialog closeDialog={closeDialog} interview={interview} />
       </Dialog>
     </div>
   );
